requirements.txt
-f /usr/share/pip-wheels
Django==1.11.3
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.0
pytz==2017.2
